+++ src/navs/index.json
@@ -344,7 +344,7 @@
             "title": "Get started",
             "classes": "lg:ml-6",
             "cta": true,
-            "url": "/sign-up",
+            "url": "/signup",
             "hideBorder": true
         },
         {
